feat(@air/zephyr/box): correctly implement "as" prop and inferred typing
Previously, we were only doing "duck" typing for the "as" prop when its value was `keyof
JSX.InstrinsicElement`. This change allows us to infer props from any component passed to the "as"
prop.

BREAKING CHANGE: This may expose typing issues that we were previously allowing in environments with
a tsconfig that allowed implicit any.