fix(@air/zephyr/modal): fix label id matching in dialog modal, tweak styles, and small refactor
For non-alert modals, if a user customizes the description, the label ID needs to still be passed to
the element. We weren't doing this before, and it's now been fixed by passing it to a headless, Box
parent. In terms of style tweaks, we crimped the height of the modal for mobile viewports and nudged
the close buton a bit more to the left so it aligns with buttons in the bottom-right corner (if they
exist). We also did some refactoring in the name of DRY and corrected references to "trap" modal
(should be "alert" modal).